chore: move artifacts into submodule
