+++ compiler/noirc_evaluator/src/ssa/mod.rs
@@ -19,10 +19,7 @@ use crate::{
 use acvm::{
     FieldElement,
     acir::{
-        circuit::{
-            AcirOpcodeLocation, Circuit, ErrorSelector, ExpressionWidth, OpcodeLocation,
-            Program as AcirProgram, PublicInputs, brillig::BrilligBytecode,
-        },
+        circuit::{AcirOpcodeLocation, Circuit, ExpressionWidth, OpcodeLocation, PublicInputs},
         native_types::Witness,
     },
 };
@@ -40,9 +37,11 @@ use tracing::{Level, span};
 
 use crate::acir::GeneratedAcir;
 
+pub use artifact::{SsaCircuitArtifact, SsaProgramArtifact};
 use builder::time;
 pub use builder::{SsaBuilder, SsaPass};
 
+mod artifact;
 mod builder;
 mod checks;
 pub mod function_builder;
@@ -333,55 +332,6 @@ where
     optimize_ssa_builder_into_acir(builder, options, primary, secondary)
 }
 
-#[derive(Default)]
-pub struct SsaProgramArtifact {
-    pub program: AcirProgram<FieldElement>,
-    pub debug: Vec<DebugInfo>,
-    pub warnings: Vec<SsaReport>,
-    pub main_input_witnesses: Vec<Witness>,
-    pub main_return_witnesses: Vec<Witness>,
-    pub names: Vec<String>,
-    pub brillig_names: Vec<String>,
-    pub error_types: BTreeMap<ErrorSelector, ErrorType>,
-}
-
-impl SsaProgramArtifact {
-    pub fn new(
-        unconstrained_functions: Vec<BrilligBytecode<FieldElement>>,
-        error_types: BTreeMap<ErrorSelector, ErrorType>,
-    ) -> Self {
-        let program = AcirProgram { functions: Vec::default(), unconstrained_functions };
-        Self {
-            program,
-            debug: Vec::default(),
-            warnings: Vec::default(),
-            main_input_witnesses: Vec::default(),
-            main_return_witnesses: Vec::default(),
-            names: Vec::default(),
-            brillig_names: Vec::default(),
-            error_types,
-        }
-    }
-
-    pub fn add_circuit(&mut self, mut circuit_artifact: SsaCircuitArtifact, is_main: bool) {
-        self.program.functions.push(circuit_artifact.circuit);
-        self.debug.push(circuit_artifact.debug_info);
-        self.warnings.append(&mut circuit_artifact.warnings);
-        if is_main {
-            self.main_input_witnesses = circuit_artifact.input_witnesses;
-            self.main_return_witnesses = circuit_artifact.return_witnesses;
-        }
-        self.names.push(circuit_artifact.name);
-        // Acir and brillig both generate new error types, so we need to merge them
-        // With the ones found during ssa generation.
-        self.error_types.extend(circuit_artifact.error_types);
-    }
-
-    fn add_warnings(&mut self, mut warnings: Vec<SsaReport>) {
-        self.warnings.append(&mut warnings);
-    }
-}
-
 /// Compiles the [`Program`] into [`ACIR`][acvm::acir::circuit::Program].
 ///
 /// The output ACIR is backend-agnostic and so must go through a transformation pass before usage in proof generation.
@@ -470,16 +420,6 @@ where
     program_artifact.brillig_names = brillig_function_names;
 
     Ok(program_artifact)
-}
-
-pub struct SsaCircuitArtifact {
-    pub name: String,
-    pub circuit: Circuit<FieldElement>,
-    pub debug_info: DebugInfo,
-    pub warnings: Vec<SsaReport>,
-    pub input_witnesses: Vec<Witness>,
-    pub return_witnesses: Vec<Witness>,
-    pub error_types: BTreeMap<ErrorSelector, ErrorType>,
 }
 
 pub fn convert_generated_acir_into_circuit(
